# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "ConnectedTo": {
        "description": "Куда подключён порт — вычисляется по наличию связи, нигде не хранится.",
        "properties": {
          "device_code": {
            "title": "Device Code",
            "type": "string"
          },
          "device_id": {
            "title": "Device Id",
            "type": "integer"
          },
          "device_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Name"
          },
          "interface_id": {
            "title": "Interface Id",
            "type": "integer"
          },
          "interface_label": {
            "title": "Interface Label",
            "type": "string"
          },
          "link_id": {
            "title": "Link Id",
            "type": "integer"
          }
        },
        "required": [
          "link_id",
          "device_id",
          "device_code",
          "interface_id",
          "interface_label"
        ],
        "title": "ConnectedTo",
        "type": "object"
      },
...
      "ConnectorTypeOut": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_cage": {
            "default": false,
            "title": "Is Cage",
            "type": "boolean"
          },
          "media": {
            "default": "copper",
            "enum": [
              "copper",
              "fiber",
              "other"
            ],
            "title": "Media",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "version": {
            "default": 1,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "name",
          "id"
        ],
        "title": "ConnectorTypeOut",
        "type": "object"
      },
...
      "FreePortOut": {
        "description": "Свободный порт для выпадающего списка подключения.",
        "properties": {
          "device_code": {
            "title": "Device Code",
            "type": "string"
          },
          "device_id": {
            "title": "Device Id",
            "type": "integer"
          },
          "device_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Name"
          },
          "device_template_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Template Name"
          },
          "interface_id": {
            "title": "Interface Id",
            "type": "integer"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "port_number": {
            "title": "Port Number",
            "type": "integer"
          }
        },
        "required": [
          "interface_id",
          "label",
          "port_number",
          "device_id",
          "device_code"
        ],
        "title": "FreePortOut",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InterfaceOut": {
        "properties": {
          "connected_to": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConnectedTo"
              },
              {
                "type": "null"
              }
            ]
          },
          "connector": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConnectorTypeOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "connector_effective": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConnectorTypeOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "device_id": {
            "title": "Device Id",
            "type": "integer"
          },
          "empty_cage": {
            "default": false,
            "title": "Empty Cage",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "ip": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ip"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "link_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Link Id"
          },
          "mac": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mac"
          },
          "mode": {
            "anyOf": [
              {
                "enum": [
                  "access",
                  "trunk",
                  "uplink"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mode"
          },
          "module": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TransceiverModuleOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "port_number": {
            "title": "Port Number",
            "type": "integer"
          },
          "trunk_vlan_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trunk Vlan Ids"
          },
          "version": {
            "default": 1,
            "title": "Version",
            "type": "integer"
          },
          "vlan_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vlan Id"
          }
        },
        "required": [
          "id",
          "device_id",
          "port_number",
          "label"
        ],
        "title": "InterfaceOut",
        "type": "object"
      },
...
      "InterfaceUpdate": {
        "description": "Правка порта у конкретного устройства.\n\nНи названия, ни номера здесь нет: они описывают саму модель техники и\nправятся в шаблоне — иначе одинаковые железки разъезжаются по названиям\nпортов. Здесь настраивается то, что у каждого экземпляра своё: адреса,\nVLAN, заметка.",
        "properties": {
          "ip": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ip"
          },
          "mac": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mac"
          },
          "mode": {
            "anyOf": [
              {
                "enum": [
                  "access",
                  "trunk",
                  "uplink"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mode"
          },
          "module_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Module Id"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "trunk_vlan_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trunk Vlan Ids"
          },
          "version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          },
          "vlan_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vlan Id"
          }
        },
        "title": "InterfaceUpdate",
        "type": "object"
      },
...
      "SearchResult": {
        "properties": {
          "device_code": {
            "title": "Device Code",
            "type": "string"
          },
          "device_id": {
            "title": "Device Id",
            "type": "integer"
          },
          "device_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Name"
          },
          "interface_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interface Id"
          },
          "interface_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interface Label"
          },
          "ip": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ip"
          },
          "mac": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mac"
          }
        },
        "required": [
          "device_id",
          "device_code"
        ],
        "title": "SearchResult",
        "type": "object"
      },
...
      "TransceiverModuleOut": {
        "properties": {
          "cage_connector_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cage Connector Id"
          },
          "connector_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Connector Id"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "version": {
            "default": 1,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "name",
          "id"
        ],
        "title": "TransceiverModuleOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "Network Documentation API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/interfaces/free": {
      "get": {
        "description": "Свободные порты — для выпадающего списка «куда воткнуть».\n\nРаньше этот список собирался в браузере из всех устройств со всеми\nпортами: чтобы предложить десяток вариантов, приходилось привезти\nдвадцать четыре тысячи. Теперь свободные порты ищет база, а сколько их\nпоказать — решает limit.\n\nЗанятым считается и порт с подвешенным кабелем: кабель в него воткнут,\nвторой его конец просто некуда включить.\n\n`device_id` сужает список до одной железки — без него на площадке с\nбольшим числом свободных портов общий (по коду устройства) список из\n`limit` записей мог не дотянуться до нужного устройства вовсе: правка\nсвязи на схеме просила «переткнуть» кабель в другой порт той же\nжелезки, а порт в выпадающем списке не появлялся.",
        "operationId": "free_interfaces_interfaces_free_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "exclude_device_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Exclude Device Id"
            }
          },
          {
            "in": "query",
            "name": "device_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Device Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-Site-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Site-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FreePortOut"
                  },
                  "title": "Response Free Interfaces Interfaces Free Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Free Interfaces",
        "tags": [
          "interfaces"
        ]
      }
    },
    "/interfaces/{interface_id}": {
      "delete": {
        "description": "Убрать порт у конкретного устройства (сняли сетевую карту).\n\nСвязь при этом не удаляется: кабель остался проложен, у него повисает\nконец — подключить его заново можно к другому порту.\nРазрешено только моделям с изменяемым составом портов.",
        "operationId": "delete_interface_interfaces__interface_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "interface_id",
            "required": true,
            "schema": {
              "title": "Interface Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-Site-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Site-Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Delete Interface",
        "tags": [
          "interfaces"
        ]
      },
      "patch": {
        "operationId": "update_interface_interfaces__interface_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "interface_id",
            "required": true,
            "schema": {
              "title": "Interface Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-Site-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Site-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InterfaceUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterfaceOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Update Interface",
        "tags": [
          "interfaces"
        ]
      }
    },
...
    "/search": {
      "get": {
        "description": "Найти устройство — по имени, коду, своему IP или MAC — либо порт —\nпо IP или MAC конкретного гнезда.\n\nРаньше здесь был один запрос, INNER JOIN от порта к устройству: он не\nвидел устройство без единого порта (медиаконвертер, ИБП), не смотрел на\n`Device.management_ip`/`Device.mac` вовсе, а совпадение по имени или\nкоду устройства фан-аутилось на одну строку на каждый порт — коммутатор\nна 24 порта одним найденным устройством съедал почти весь лимит выдачи.\n\nТеперь два независимых запроса: устройство находится по своим полям —\nодной строкой, порт — по своим, отдельной строкой на каждое совпадение.\nСтрока устройства порта не называет (`interface_id`/`interface_label`\nпусты) — искали не гнездо, показывать какое-то одно было бы обманом.",
        "operationId": "search_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "required": true,
            "schema": {
              "title": "Query",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Site-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Site-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SearchResult"
                  },
                  "title": "Response Search Search Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Search",
        "tags": [
          "interfaces"
        ]
      }
    },
...
  }
}